@@ -67,18 +67,29 @@
     "typescript": "4.2.2"
   },
   "engines": {
-    "node": "^8.13.0 || >=10.10.0"
+    "node": ">=10.10.0"
   },
   "esm5": "dist/index.esm5.js",
   "exports": {
     ".": {
-      "node": "./dist/index.node.cjs.js",
+      "node": {
+        "require": "./dist/index.node.cjs.js",
+        "import": "./dist/index.node.mjs"
+      },
+      "react-native": "./dist/index.rn.js",
+      "esm5": "./dist/index.esm5.js",
       "default": "./dist/index.esm2017.js"
     },
     "./lite": {
-      "node": "./dist/lite/index.node.cjs.js",
+      "node": {
+        "require": "./dist/lite/index.node.cjs.js",
+        "import": "./dist/lite/index.node.mjs"
+      },
+      "react-native": "./dist/lite/index.rn.esm2017.js",
+      "esm5": "./dist/lite/index.browser.esm5.js",
       "default": "./dist/lite/index.browser.esm2017.js"
-    }
+    },
+    "./package.json": "./package.json"
   },
   "files": [
     "dist",
