@@ -1,8 +1,8 @@
 {
   "_from": "ws@>=7.4.6",
-  "_id": "ws@8.2.2",
+  "_id": "ws@8.3.0",
   "_inBundle": false,
-  "_integrity": "sha512-Q6B6H2oc8QY3llc3cB8kVmQ6pnJWVQbP7Q5algTcIxx7YEpc0oU4NBVHlztA7Ekzfhw2r0rPducMUiCGWKQRzw==",
+  "_integrity": "sha512-Gs5EZtpqZzLvmIM59w4igITU57lrtYVFneaa434VROv4thzJyV6UjIL3D42lslWlI+D4KzLYnxSwtfuiO79sNw==",
   "_location": "/selenium-webdriver/ws",
   "_phantomChildren": {},
   "_requested": {
@@ -18,8 +18,8 @@
   "_requiredBy": [
     "/selenium-webdriver"
   ],
-  "_resolved": "https://registry.npmjs.org/ws/-/ws-8.2.2.tgz",
-  "_shasum": "ca684330c6dd6076a737250ed81ac1606cb0a63e",
+  "_resolved": "https://registry.npmjs.org/ws/-/ws-8.3.0.tgz",
+  "_shasum": "7185e252c8973a60d57170175ff55fdbd116070d",
   "_spec": "ws@>=7.4.6",
   "_where": "C:\\Users\\leona\\Documents\\Docs1\\JML_Musculacao\\JML_Musculacao\\frontend\\node_modules\\selenium-webdriver",
   "author": {
@@ -37,7 +37,7 @@
   "devDependencies": {
     "benchmark": "^2.1.4",
     "bufferutil": "^4.0.1",
-    "eslint": "^7.2.0",
+    "eslint": "^8.0.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-prettier": "^4.0.0",
     "mocha": "^8.4.0",
@@ -91,5 +91,5 @@
     "lint": "eslint --ignore-path .gitignore . && prettier --check --ignore-path .gitignore \"**/*.{json,md,yaml,yml}\"",
     "test": "nyc --reporter=lcov --reporter=text mocha --throw-deprecation test/*.test.js"
   },
-  "version": "8.2.2"
+  "version": "8.3.0"
 }
